+++ docs/apps/schedy/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 ### Changed
+* Lowered delay after which a schedy_reschedule event is processed from
+  3 to 1 second.
 
 ### Deprecated